Pepper comments throughout the app
