+++ js/reducers/plugins.js
@@ -1,10 +1,13 @@
-import { LOAD_PLUGINS_SUCCESS } from "../actions/Plugins";
+/**
+ * Plugin reducer
+ */
+import { LOAD_PLUGINS_SUCCESS } from "../actions";
 
 const INITIAL_STATE = [];
 
 const pluginsReducer = (state = INITIAL_STATE, action) => {
     switch (action.type) {
-        case LOAD_PLUGINS_SUCCESS: 
+        case LOAD_PLUGINS_SUCCESS:
             return action.plugins;
         default:
             return state;
