--- js/tags.js
@@ -1,21 +1,0 @@
-const Tags = [
-		"analysis",
-		"color",
-		"debug",
-		"extensions",
-		"fallbacks",
-		"fonts",
-		"future",
-		"fun",
-		"grids",
-		"images",
-		"media-queries",
-		"optimizations",
-		"other",
-		"pack",
-		"sass",
-		"shortcuts",
-		"svg"
-	];
-    
-export default Tags;
