+++ js/utils/url.js
@@ -1,4 +1,6 @@
-
+/**
+ * Get the tag name from the URL
+ */
 const getTag = (pathname) => {
   const locationPieces = pathname.substr(1).split('/');
   const tag = locationPieces.length === 2 && locationPieces[0] === 'tag' ? (locationPieces[1]) : '';
